+++ examples/db_remember.rs
@@ -1,12 +1,23 @@
-use teloxide::{macros::DialogueState, prelude2::*, types::Me, utils::command::BotCommand};
+// Set the `DB_REMEMBER_REDIS` environmental variable if you want to use Redis.
+// Otherwise, the default is Sqlite.
 
-// include!("redis_config.in");
-include!("sqlite_config.in");
+use teloxide::{
+    dispatching2::dialogue::{
+        serializer::{Bincode, Json},
+        ErasedStorage, RedisStorage, SqliteStorage, Storage,
+    },
+    macros::DialogueState,
+    prelude2::*,
+    types::Me,
+    utils::command::BotCommand,
+};
 
-type MyDialogue = Dialogue<State, MyStorage>;
+type MyDialogue = Dialogue<State, ErasedStorage<State>>;
+type MyStorage = std::sync::Arc<ErasedStorage<State>>;
+type HandlerResult = Result<(), Box<dyn std::error::Error + Send + Sync>>;
 
 #[derive(DialogueState, Clone, serde::Serialize, serde::Deserialize)]
-#[handler_out(anyhow::Result<()>)]
+#[handler_out(HandlerResult)]
 pub enum State {
     #[handler(handle_start)]
     Start,
@@ -34,9 +45,14 @@ pub enum Command {
 async fn main() {
     let bot = Bot::from_env().auto_send();
 
-    let storage = open_storage().await;
+    let storage: MyStorage = if std::env::var("DB_REMEMBER_REDIS").is_ok() {
+        RedisStorage::open("redis://127.0.0.1:6379", Bincode).await.unwrap().erase()
+    } else {
+        SqliteStorage::open("db.sqlite", Json).await.unwrap().erase()
+    };
+
     let handler = Update::filter_message()
-        .enter_dialogue::<Message, MyStorage, State>()
+        .enter_dialogue::<Message, ErasedStorage<State>, State>()
         .dispatch_by::<State>();
 
     Dispatcher::builder(bot, handler)
@@ -47,11 +63,7 @@ async fn main() {
         .await;
 }
 
-async fn handle_start(
-    bot: AutoSend<Bot>,
-    msg: Message,
-    dialogue: MyDialogue,
-) -> anyhow::Result<()> {
+async fn handle_start(bot: AutoSend<Bot>, msg: Message, dialogue: MyDialogue) -> HandlerResult {
     match msg.text().unwrap().parse() {
         Ok(number) => {
             dialogue.update(State::GotNumber(number)).await?;
@@ -75,7 +87,7 @@ async fn handle_got_number(
     dialogue: MyDialogue,
     num: i32,
     me: Me,
-) -> anyhow::Result<()> {
+) -> HandlerResult {
     let ans = msg.text().unwrap();
     let bot_name = me.user.username.unwrap();