+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The `Storage::erase` default function that returns `ErasedStorage`.
+- The `Storage::erase` default function that returns `Arc<ErasedStorage>`.
 - `ErasedStorage`, a storage with an erased error type.
+- Allow the storage generic `S` be `?Sized` in `Dialogue`.
 
 ## 0.7.1 - 2022-03-09