+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `Storage::erase` default function that returns `ErasedStorage`.
 - `ErasedStorage`, a storage with an erased error type.
 
+### Fixed
+
+- Log `UpdateKind::Error` in `teloxide::dispatching2::Dispatcher`.
+
 ## 0.7.1 - 2022-03-09
 
 ### Fixed