+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.4.4 - 2022-04-21
+
 ### Added
 
 - `WrongFileIdOrUrl` and `FailedToGetUrlContent` errors ([#188][pr188])